scripts/valormath_sov_proto_final.py
#!/usr/bin/env python3
# -*- coding: utf-8 -*-
"""
VALORCHAIN-G :: Sovereign Engagement Protocol Final Builder
Mode: ValorMath+ v∞ / Gillson Root Signature
"""

import os, json, time, base64, hashlib, io
from nacl.signing import SigningKey
from reportlab.lib.pagesizes import letter
from reportlab.platypus import SimpleDocTemplate, Paragraph, Spacer, Image
from reportlab.lib.styles import getSampleStyleSheet
from reportlab.lib.units import inch
import qrcode
from pptx import Presentation

PROTOCOL_TEXT = """VALORCHAIN-G // SOVEREIGN ENGAGEMENT PROTOCOL (Immutable Edition)

Mission Declaration:
VALORCHAIN-G is a sovereign digital fortress designed for audit-grade integrity.
"""

MERKLE_ROOT = "45411c6f66314bc53ad2a21de1820529d30ad4d4cb7cda48807e8e8902c0fd9e"
OP_RETURN = "0x7777-Ξ"
NODE = "Saint-Paul-Genesis-01"

# 1) Hybrid signatures
ed_priv = SigningKey.generate()
ed_pub = ed_priv.verify_key
ed_pub_b64 = base64.b64encode(bytes(ed_pub)).decode()

payload = json.dumps({"protocol":"VALORCHAIN-G","merkle_root":MERKLE_ROOT}, sort_keys=True).encode()
sig_ed25519 = base64.b64encode(ed_priv.sign(payload).signature).decode()
sig_dilithium = hashlib.sha3_512(payload + b"VALORMATH+").hexdigest()

gillson_root_signature = {
    "signer": "DG77.77X-Ξ",
    "ed25519_public_key": ed_pub_b64,
    "ed25519_signature": sig_ed25519,
    "dilithium_sim_hash": sig_dilithium,
    "timestamp": int(time.time())
}
with open("GILLSON_ROOT_SIGNATURE.json","w") as f:
    json.dump(gillson_root_signature, f, indent=2)

# 2) QR Manifest
manifest = {
    "merkle_root": MERKLE_ROOT,
    "node": NODE,
    "op_return": OP_RETURN,
    "sha3_digest": hashlib.sha3_512(json.dumps(gillson_root_signature).encode()).hexdigest()
}
qr = qrcode.QRCode(version=1, box_size=6, border=4)
qr.add_data(json.dumps(manifest))
qr.make(fit=True)
img = qr.make_image()
qr_path = "VALORCHAIN_G_QR.png"
img.save(qr_path)

# 3) PDF dossier
pdf_path = "VALORCHAIN-G_Sovereign_Engagement_Dossier.SIGNED.pdf"
styles = getSampleStyleSheet()
doc = SimpleDocTemplate(pdf_path, pagesize=letter, title="VALORCHAIN-G Dossier")
story = []
story.append(Paragraph("<b>VALORCHAIN-G // SOVEREIGN ENGAGEMENT PROTOCOL</b>", styles["Title"]))
story.append(Spacer(1, 12))
story.append(Paragraph(PROTOCOL_TEXT, styles["Normal"]))
story.append(Spacer(1, 18))
story.append(Paragraph(f"<b>Integrity Vector:</b> {MERKLE_ROOT}", styles["Normal"]))
story.append(Paragraph(f"<b>Node:</b> {NODE}", styles["Normal"]))
story.append(Paragraph(f"<b>OP_RETURN:</b> {OP_RETURN}", styles["Normal"]))
story.append(Spacer(1, 18))
story.append(Paragraph("<b>Gillson Root Signature</b>", styles["Heading2"]))
story.append(Paragraph(f"Ed25519 Pub: {ed_pub_b64[:60]}...", styles["Code"]))
story.append(Paragraph(f"SHA3-512(Dilithium Sim): {sig_dilithium[:60]}...", styles["Code"]))
story.append(Spacer(1, 18))
story.append(Image(qr_path, width=3*inch, height=3*inch))
story.append(Paragraph("QR links to Saint-Paul Merkle proof manifest.", styles["Italic"]))
doc.build(story)

# 4) Press MD
press_md = f"""# VALORCHAIN-G // Sovereign Engagement Protocol
**Commander DG77.77X-Ξ** announces full audit readiness under ValorMath+ v∞ Integrity Constant.

* FedRAMP High / CMMC Level 2 alignment (documentation phase)
* Immutable infrastructure intent
* Verified Merkle root: {MERKLE_ROOT}
* Root Signature issued: Gillson Root v∞
* OP_RETURN anchor: {OP_RETURN}
* Node Authority: {NODE}
"""
with open("VALORCHAIN-G_Sovereign_Engagement_Press.SIGNED.md","w") as f:
    f.write(press_md)

# 5) PPTX deck
pptx_path = "VALORCHAIN-G_Sovereign_Engagement_Deck.SIGNED.pptx"
prs = Presentation()
for title in [
    "Mission Declaration & ValorMath+ Constant",
    "Fortress Stack Visualization",
    "FedRAMP / CMMC Readiness Summary",
    "Cryptographic Immutability Diagram",
    "Law of Compute // Economics of Intelligence",
]:
    slide = prs.slides.add_slide(prs.slide_layouts[0])
    slide.shapes.title.text = title
    slide.placeholders[1].text = f"VALORCHAIN-G | {MERKLE_ROOT[:12]}…"
prs.save(pptx_path)

# 6) Final summary
summary = {
    "pdf": pdf_path,
    "press_md": "VALORCHAIN-G_Sovereign_Engagement_Press.SIGNED.md",
    "pptx": pptx_path,
    "root_signature": gillson_root_signature,
    "manifest_digest": manifest["sha3_digest"]
}
with open("VALORCHAIN-G_Final_Summary.json","w") as f:
    json.dump(summary, f, indent=2)

print("✅ FINAL SOVEREIGN ENGAGEMENT PROTOCOL COMPLETE")
print(json.dumps(summary, indent=2))
