requirements.txt
pynacl
reportlab
python-pptx
qrcode[pil]
